[OGR provider] Handles layer of geometry type unknown with null geometries
Fixes #50215

Layers with a wkbUnknown OGR layer geometry types are first scanned to
figure out which actualy geometry types are present, to expose sublayers
filtered with the |geometryType=xxxx URI parameter.

That hint is then used by getFeatureCount() to only count features with
such geometry type, and also by the feature iterator. A consequence of
that is that features with null geometries are excluded.

When there is a single geometry type, we could then just use the general
purpose code paths, ie not actually filter on geometry type, so do that
by adding a |uniqueGeometryType=yes URI parameter.
Note that this may not be fully bullet-proof if adding such layer to a
project, and later the datasource is modified to add geometries of
another type.

And note also that there is another issue, pre-existing to this PR and
not addressed by it, if a layer contains a feature with a geometry of
type A, one of type B and a feature with no geometry.
QGIS will never report the feature with no geometry.

Related GDAL fix specifically for the original use case of #50215 in
https://github.com/OSGeo/gdal/pull/6377