Make "archive" copy of QgsPalLayerSettings::calculateLabelSize
For stable API use only, and then cleanup the private version
of this to remove all the consideration of 3.0 API compatibility.

Mark the archive copy as deprecated and remove in 4.0

This method has become completely unweildy because of all the
compatibility code. By making this private only we can rip
out all this old code and make the method much simpler and
easy to understand.