Use optimised approach for drive volume name retrieval on windows
QStorageInfo can be VERY slow on windows, if the drive is a remote
device, as the QStorageInfo constructor FORCES calculation
of drive available space.

Since we never use that and instead only require the volume name,
directly use Windows API to retrieve ONLY this.

Fixes slow app opening on Windows with slow network drives.